requirements.txt
# Core dependencies
Telethon>=1.36.0
python-dotenv>=1.0.0

# Telethon dependencies (installed automatically)
# pyaes
# pyasn1
# rsa
